finished refactoring filterCards
