+++ src/Components/MuniFilterCard.js
@@ -1,5 +1,4 @@
 import { useState } from 'react';
-
 import '@fontsource/roboto';
 import {
 	Card,
